pyproject.toml
[tool.poetry]
name = "orm"
version = "0.1.0"
description = ""
authors = ["Eugene Morozov <jmv@emorozov.net>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.6"
ipython = "^8.4.0"
django-extensions = "^3.2.0"

[tool.poetry.dev-dependencies]
factory-boy = "^3.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
